Map DOUBLE to double.class, not float.class

+++ src/main/java/jnr/ffi/provider/jffi/X86MethodGenerator.java
@@ -354,7 +354,7 @@ class X86MethodGenerator implements MethodGenerator {
                 return float.class;
 
             case DOUBLE:
-                return float.class;
+                return double.class;
 
             case VOID:
                 return void.class;
